usuarios y pisos mejorado

+++ backend/eliminar_piso.php
@@ -6,8 +6,22 @@ if ($_SESSION['tipo'] != 'admin') {
     header("Location: ../login.php");
     exit();
 }
+// VALIDAR ID
+if (!isset($_GET['id'])) {
+    header("Location: pisos.php");
+    exit();
+}
 
-$id = $_GET['id'];
+
+$id = intval($_GET['id']);
+
+$sql_check = "SELECT * FROM pisos WHERE id=$id";
+$result = $conn->query($sql_check);
+
+if ($result->num_rows == 0) {
+    echo "Piso no encontrado";
+    exit();
+}
 
 $sql = "DELETE FROM pisos WHERE id=$id";
